Move `TiersComponentView` to separate file
